# order-service/api/openapi.yaml
openapi: 3.0.3
info:
  title: Order Service
  version: 1.0.0
  description: API для управления заказами

paths:
  /api/v1/orders:
    post:
      operationId: createOrder
      summary: Создать новый заказ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "400":
          description: Неверный запрос (невалидные данные)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Заказ с таким UUID уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Неожиданная ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/orders/{order_uuid}:
    get:
      operationId: getOrder
      summary: Получить информацию о заказе
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            type: string

      responses:
        "200":
          description: Заказ найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Неожиданная ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/orders/{order_uuid}/pay:
    post:
      operationId: payOrder
      summary: Оплатить заказ
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            type: string

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayOrderRequest"
      responses:
        "200":
          description: Заказ успешно оплачен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayOrderResponse"
        "400":
          description: Неверный метод оплаты или данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Заказ уже оплачен или в неверном статусе
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Ошибка при оплате
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Неожиданная ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/orders/{order_uuid}/cancel:
    post:
      operationId: cancelOrder
      summary: Отменить заказ
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            type: string

      responses:
        "204":
          description: Заказ успешно отменён (без тела ответа)
        "404":
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Заказ уже оплачен и не может быть отменён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Неожиданная ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: "заказ не найден"
        code:
          type: string
          example: "NOT_FOUND"
      required:
        - message

    Order:
      type: object
      required:
        - order_uuid
        - user_uuid
        - part_uuids
        - total_price
        - status
      properties:
        order_uuid:
          type: string

        user_uuid:
          type: string
        part_uuids:
          type: array
          items:
            type: string
        total_price:
          type: number
          format: double
        transaction_uuid:
          type: string
          nullable: true
        payment_method:
          type: string
          nullable: true
          enum: [CARD, SBP, CREDIT_CARD, INVESTOR_MONEY]
        status:
          type: string
          enum: [PENDING_PAYMENT, PAID, CANCELLED]

    CreateOrderRequest:
      type: object
      required: [user_uuid, part_uuids]
      properties:
        user_uuid:
          type: string
        part_uuids:
          type: array
          items:
            type: string

    CreateOrderResponse:
      type: object
      required: [order_uuid, total_price]
      properties:
        order_uuid:
          type: string

        total_price:
          type: number
          format: double

    PayOrderRequest:
      type: object
      required: [payment_method]
      properties:
        payment_method:
          type: string
          enum: [CARD, SBP, CREDIT_CARD, INVESTOR_MONEY]

    PayOrderResponse:
      type: object
      required: [transaction_uuid]
      properties:
        transaction_uuid:
          type: string